Modify readme and rename packages
