+++ src/main/java/sariel/util/commands/AddCommand.java
@@ -1,9 +1,9 @@
-package duke.util.commands;
+package sariel.util.commands;
 
-import duke.util.commons.Messages;
-import duke.util.tasks.Task;
-import duke.util.tasks.TaskList;
-import duke.util.ui.Ui;
+import sariel.util.commons.Messages;
+import sariel.util.tasks.Task;
+import sariel.util.tasks.TaskList;
+import sariel.util.ui.Ui;
 
 /**
  * Command representing the addition of a
