+++ src/main/java/sariel/util/commands/ExitCommand.java
@@ -1,4 +1,4 @@
-package duke.util.commands;
+package sariel.util.commands;
 
 
 /**
