+++ src/main/java/sariel/util/controller/Sariel.java
@@ -1,24 +1,24 @@
-package duke.util.controller;
+package sariel.util.controller;
 
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.function.Predicate;
 
-import duke.util.commands.AddCommand;
-import duke.util.commands.CommandList;
-import duke.util.commons.Messages;
-import duke.util.parser.Parser;
-import duke.util.storage.Storage;
-import duke.util.tasks.DateTaskTable;
-import duke.util.tasks.Deadline;
-import duke.util.tasks.DukeException;
-import duke.util.tasks.Event;
-import duke.util.tasks.Task;
-import duke.util.tasks.TaskList;
-import duke.util.tasks.ToDo;
-import duke.util.ui.Ui;
-import duke.util.windows.DialogBox;
+import sariel.util.commands.AddCommand;
+import sariel.util.commands.CommandList;
+import sariel.util.commons.Messages;
+import sariel.util.parser.Parser;
+import sariel.util.storage.Storage;
+import sariel.util.tasks.DateTaskTable;
+import sariel.util.tasks.Deadline;
+import sariel.util.tasks.DukeException;
+import sariel.util.tasks.Event;
+import sariel.util.tasks.Task;
+import sariel.util.tasks.TaskList;
+import sariel.util.tasks.ToDo;
+import sariel.util.ui.Ui;
+import sariel.util.windows.DialogBox;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.control.ListView;
