+++ src/main/java/sariel/util/parser/Parser.java
@@ -1,25 +1,25 @@
-package duke.util.parser;
+package sariel.util.parser;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.function.Predicate;
 
-import duke.util.commands.AddCommand;
-import duke.util.commands.CommandList;
-import duke.util.commands.DelCommand;
-import duke.util.commands.DoneCommand;
-import duke.util.commands.ExitCommand;
-import duke.util.commons.Messages;
-import duke.util.tasks.DateTaskTable;
-import duke.util.tasks.DatedTask;
-import duke.util.tasks.Deadline;
-import duke.util.tasks.DukeException;
-import duke.util.tasks.Event;
-import duke.util.tasks.Task;
-import duke.util.tasks.TaskList;
-import duke.util.tasks.ToDo;
-import duke.util.ui.Ui;
+import sariel.util.commands.AddCommand;
+import sariel.util.commands.CommandList;
+import sariel.util.commands.DelCommand;
+import sariel.util.commands.DoneCommand;
+import sariel.util.commands.ExitCommand;
+import sariel.util.commons.Messages;
+import sariel.util.tasks.DateTaskTable;
+import sariel.util.tasks.DatedTask;
+import sariel.util.tasks.Deadline;
+import sariel.util.tasks.DukeException;
+import sariel.util.tasks.Event;
+import sariel.util.tasks.Task;
+import sariel.util.tasks.TaskList;
+import sariel.util.tasks.ToDo;
+import sariel.util.ui.Ui;
