+++ src/main/java/sariel/util/tasks/Event.java
@@ -1,6 +1,6 @@
-package duke.util.tasks;
+package sariel.util.tasks;
 
-import duke.util.commons.Messages;
+import sariel.util.commons.Messages;
 
 public class Event extends DatedTask {
     private static String label = "[E]";
