+++ src/main/java/sariel/util/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package duke.util.tasks;
+package sariel.util.tasks;
 
 public class ToDo extends Task {
