+++ src/main/java/sariel/util/windows/MainWindow.java
@@ -1,10 +1,10 @@
-package duke.util.windows;
+package sariel.util.windows;
 
 import java.io.IOException;
 
-import duke.util.commons.Messages;
-import duke.util.controller.Sariel;
-import duke.util.tasks.Task;
+import sariel.util.commons.Messages;
+import sariel.util.controller.Sariel;
+import sariel.util.tasks.Task;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
