+++ src/main/java/sariel/util/windows/TaskInputWindow.java
@@ -1,9 +1,9 @@
-package duke.util.windows;
+package sariel.util.windows;
 
 import java.time.LocalDate;
 
-import duke.util.commons.Messages;
-import duke.util.controller.Sariel;
+import sariel.util.commons.Messages;
+import sariel.util.controller.Sariel;
 import javafx.fxml.FXML;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.Tab;
